fix full page cache issue
