+++ etc/module.xml
@@ -1,5 +1,9 @@
 <?xml version="1.0"?>
 
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:Module/etc/module.xsd">
-    <module name="Chapagain_AutoCurrency" setup_version="2.1.1" />
+    <module name="Chapagain_AutoCurrency" setup_version="2.1.2">
+        <sequence>
+            <module name="Magento_PageCache"/>
+        </sequence>
+    </module>
 </config>
